Upload Image and get prediction
Upload an image to the server for model inference.

The server saves the image, does inference on the image, and hosts the
image for the client to view.

Things to note: This solution stores the images on the FS instead of in
a database or similar. I dont want to pay for a website that allows
image uploads. Solution assumes model is in App directory and named
'export.pkl'.